fix to issue #1  - Import not working
the issue is that the current import was attempting to parse the
category id of the FAQ. However if the FAQ is not in a category the code
threw an exception. This fix will check to see if we have a category
first defined on the faq item in the import xml.